+++ pandas/tests/indexes/datetimes/test_ops.py
@@ -45,11 +45,13 @@ class TestDatetimeIndexVectorizedTimestamp(object):
         assert idx.freqstr == Timestamp(idx[-1], idx.freq).freqstr
 
     def test_round(self):
+        # tz-naive
         dti = date_range('20130101 09:10:11', periods=5)
         result = dti.round('D')
         expected = date_range('20130101', periods=5)
         tm.assert_index_equal(result, expected)
 
+        # tz-aware
         dti = date_range('20130101 09:10:11',
                          periods=5).tz_localize('UTC').tz_convert('US/Eastern')
         result = dti.round('D')
